Add user types document
This is an initial start for the various types of users who will be interacting with our system. This is not an exhaustive list, nor is any user fully documented. Most of the information was pulled from the Challenge Post listing.

The intention of this document is to have a place to start defining our users.